+++ src/hooks/useDetailMap.js
@@ -1,13 +1,15 @@
 import { useQuery } from '@tanstack/react-query';
-import supabase from '../supabase/supabase';
+
 import Swal from 'sweetalert2';
+
+import supabase from '../supabase/supabase';
 
 const useDetailMap = (id) => {
   // 해당 식당 주소, 이름, 설명을 가져온다.
   const getRestorantAddress = async (restaurantId) => {
     const { data, error } = await supabase
       .from('restaurants')
-      .select('name, address, description')
+      .select('name, address, description, phone_number')
       .eq('id', restaurantId)
       .single();
