+++ src/index.css
@@ -32,3 +32,14 @@
   font-size: 14px;
   color: #8f8f8f;
 }
+
+.phone-number {
+  margin-top: 5px;
+  color: #8f8f8f;
+  transition: all ease-in 0.2s;
+  font-size: 14px;
+}
+
+.phone-number:hover {
+  color: black;
+}
